chore: Update GBAN API endpoint to production URL
This commit updates the API endpoint for the GBAN script to the production URL "http://api.gban.un-known.xyz/v1/search". The previous URL "http://localhost/v1/search" was used for testing purposes and is no longer needed. This change ensures that the script is using the correct endpoint in the production environment.